Update README.md
Submitting PR to improve the Readme. The line which says:
"The higher valued index should be placed in the `zeroth` index and the smaller index should be placed in the `first` index". 

Recommend dropping the word 'valued' to make this clearer as it reads as if the higher value of the index instead of the higher index.
Multiple students commented on this as confusing.
Also since the index is the value in dictionary, it further complicates the intent so the proposed tweak would make this clearer.